refactor(renderer): extract launcher command model
